# admin-panel/generated/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Shortlinker API",
    "description": "Shortlinker management API",
    "contact": {
      "name": "AptS-1547",
      "email": "apts-1547@esaps.net"
    },
    "license": {
      "name": "MIT"
    },
    "version": "0.5.2"
  },
  "paths": {
...
    "/admin/v1/links/batch": {
      "put": {
        "tags": [
          "links"
        ],
        "summary": "批量更新链接",
        "operationId": "batch_update_links",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BatchUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Batch update result",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_BatchResponse"
                }
              }
            }
          },
          "400": {
            "description": "Batch too large or invalid"
          }
        }
      },
      "post": {
        "tags": [
          "links"
        ],
        "summary": "批量创建链接",
        "operationId": "batch_create_links",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BatchCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Batch create result",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_BatchResponse"
                }
              }
            }
          },
          "400": {
            "description": "Batch too large or invalid"
          }
        }
      },
      "delete": {
        "tags": [
          "links"
        ],
        "summary": "批量删除链接",
        "operationId": "batch_delete_links",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BatchDeleteRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Batch delete result",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_BatchResponse"
                }
              }
            }
          },
          "400": {
            "description": "Batch too large or invalid"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ApiResponse_BatchResponse": {
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "properties": {
          "code": {
            "type": "integer",
            "format": "int32"
          },
          "data": {
            "type": "object",
            "required": [
              "success",
              "failed"
            ],
            "properties": {
              "failed": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/BatchFailedItem"
                }
              },
              "success": {
                "type": "array",
                "items": {
                  "type": "string"
                }
              }
            }
          },
          "message": {
            "type": "string"
          }
        }
      },
...
      "BatchCreateRequest": {
        "type": "object",
        "required": [
          "links"
        ],
        "properties": {
          "links": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PostNewLink"
            }
          }
        }
      },
      "BatchDeleteRequest": {
        "type": "object",
        "required": [
          "codes"
        ],
        "properties": {
          "codes": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
      "BatchFailedItem": {
        "type": "object",
        "required": [
          "code",
          "error"
        ],
        "properties": {
          "code": {
            "type": "string"
          },
          "error": {
            "type": "string"
          },
          "error_code": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32"
          }
        }
      },
...
      "BatchUpdateItem": {
        "type": "object",
        "required": [
          "code",
          "payload"
        ],
        "properties": {
          "code": {
            "type": "string"
          },
          "payload": {
            "$ref": "#/components/schemas/PostNewLink"
          }
        }
      },
      "BatchUpdateRequest": {
        "type": "object",
        "required": [
          "updates"
        ],
        "properties": {
          "updates": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/BatchUpdateItem"
            }
          }
        }
      },
...
      "PostNewLink": {
        "type": "object",
        "required": [
          "target"
        ],
        "properties": {
          "code": {
            "type": [
              "string",
              "null"
            ]
          },
          "expires_at": {
            "type": [
              "string",
              "null"
            ]
          },
          "force": {
            "type": [
              "boolean",
              "null"
            ]
          },
          "password": {
            "type": [
              "string",
              "null"
            ]
          },
          "target": {
            "type": "string"
          }
        }
      },
...
    }
  },
...
}